address CI: move ratchet to test_doc_invariants.py; add changelog entry
The meta-ratchet test_all_test_ratchets_files_have_same_tests requires
test_ratchets.py to contain exactly the standard ratchet set (plus
test_no_type_errors / test_no_ruff_errors). The mngr-specific
test_enum_matches_docs_table didn't belong there. Move it to a new
libs/mngr/imbue/mngr/utils/test_doc_invariants.py, restore test_ratchets.py
to its origin/main state, and update the style_guide.md pointer.

Also add changelog/mngr-d614.md so test_pr_has_changelog_entry passes.
