+++ changelog/mngr-d614.md
@@ -1,0 +1,1 @@
+Three docs pages (`docs/concepts/hosts.md`, `docs/concepts/agents.md`, `docs/concepts/idle_detection.md`) now list every member of their respective enums (`HostState`, `AgentLifecycleState`, `IdleMode`); the new ratchet `test_enum_matches_docs_table` enforces this invariant going forward. Style guide gains a "Programmatic enforcement of doc-vs-code invariants" section that explains the rationale and points at the new ratchet as the canonical example.
